consolidate lexer into 1 file
